app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
pillow
transformers>=4.0.0
requests>=2.27.1
numpy
imageio
